make test_base_passage_compressor.py to prevent circular import error.
