+++ tests/autorag/nodes/passagecompressor/test_tree_summarize.py
@@ -1,27 +1,10 @@
-from typing import List
-
 import pandas as pd
 
 from autorag import generator_models
 from autorag.nodes.passagecompressor import tree_summarize
+from tests.autorag.nodes.passagecompressor.test_base_passage_compressor import (queries, retrieved_contents, df,
+                                                                                check_result)
 from tests.mock import MockLLM
-
-queries = [
-    "What is the capital of France?",
-    "What is the meaning of life?",
-]
-retrieved_contents = [
-    ["Paris is the capital of France.", "France is a country in Europe.", "France is a member of the EU."],
-    ["The meaning of life is 42.", "The meaning of life is to be happy.", "The meaning of life is to be kind."],
-]
-
-
-def check_result(result: List[str]):
-    assert len(result) == len(queries)
-    for r in result:
-        assert isinstance(r, str)
-        assert len(r) > 0
-        assert bool(r) is True
 
 
 def test_tree_summarize_default():
@@ -50,14 +33,6 @@ def test_tree_summarize_custom_prompt_chat():
     result = tree_summarize.__wrapped__(queries, retrieved_contents, [], [], gpt_3, chat_prompt=prompt)
     check_result(result)
     assert 'What is the capital of France?' in result[0]
-
-
-df = pd.DataFrame({
-        'query': queries,
-        'retrieved_contents': retrieved_contents,
-        'retrieved_ids': [['id-1', 'id-2', 'id-3'], ['id-4', 'id-5', 'id-6']],
-        'retrieve_scores': [[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]],
-    })
 
 
 def test_tree_summarize_node():
